# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /paths/{pathId}:
      x-swagger-router-controller: paths
      get:
        operationId: show
        summary: A learning path
        tags:
          - Path
        parameters:
          - name: pathId
            in: path
            required: true
            type: string
        responses:
          200:
            description: successful operation
            schema:
              $ref: "#/definitions/Path"
      put:
        operationId: update
        summary: Update path
        tags:
          - Path
        parameters:
          - name: pathId
            in: path
            required: true
            type: string
          - name: Path
            in: body
            required: true
            schema:
              $ref: "#/definitions/Path"
        responses:
          204:
            description: path updated
          400:
            description: Invalid input
      delete:
        operationId: destroy
        summary: Delete path
        tags:
          - Path
        parameters:
          - name: pathId
            in: path
            required: true
            type: string
        responses:
          204:
            description: path deleted
...
  /categories/{categoryId}:
    x-swagger-router-controller: categories
    get:
      operationId: show
      summary: A category name
      tags:
        - Categories
      parameters:
        - name: categoryId
          in: path
          required: true
          type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: "#/definitions/Category"

    delete:
      operationId: destroy
      summary: Delete a category from the list of categories (admin access required)
      tags:
        - Categories
      parameters:
        - name: categoryId
          in: path
          required: true
          type: string
      responses:
        204:
          description: successful operation

    put:
      operationId: update
      summary: Update a category in the list of categories (admin access required)
      tags:
        - Categories
      parameters:
        - name: categoryId
          in: path
          required: true
          type: string
        - name: Category
          in: body
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        204:
          description: category updated
        400:
          description: Invalid input
...
definitions:
...
  Path:
    type: object
    properties:
      _id:
        type: string
      categories:
        type: array
        items:
          $ref: "#/definitions/Category"
      name:
        type: string
        description: Name of the learning path
      description:
        type: string
        description: Short description of the learning path
      author:
        $ref: "#/definitions/User"
      steps:
        type: array
        items:
          $ref: "#/definitions/PathStep"
      comments:
        type: array
        items:
          $ref: "#/definitions/PathComment"
      individual:
        type: boolean
      group:
        type: boolean
      rating:
        type: integer
        minimum: 1
        maximum: 5
      admin:
        $ref: "#/definitions/User"
      participants:
        type: array
        items:
          $ref: "#/definitions/User"
      originalPathId:
        type: string
        description: id of the original path. If this is an empty string, the path is original, else it is a clone
      clonedBy:
        type: array
        items:
          $ref: "#/definitions/User"
      fromDateOfEnrolment:
        type: string
        format: date-time
      toDateOfEnrolment:
        type: string
        format: date-time
  Category:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        slug:
          type: string
  User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
          description: Name of the user
        avatar:
          type: string
          description: URL of the user's image
        pathsAuthored:
          type: number
          format: int32
        pathsCloned:
          type: number
          format: int32
  PathStep:
    type: object
    properties:
      _id:
        type: string
      name:
        type: string
      deadline:
        type: string
        format: date-time
      completed:
        type: boolean
      description:
        type: string
      notes:
        type: array
        items:
          $ref: "#/definitions/StepNote"
  PathComment:
    type: object
    properties:
      _id:
        type: string
      author:
        $ref: "#/definitions/User"
      pathId:
        type: string
      parentId:
        type: string
      slug:
        type: string
      fullslug:
        type: string
      posted:
        type: string
        format: date-time
      text:
        type: string
  StepNote:
    type: object
    properties:
      _id:
        type: string
      author:
        $ref: "#/definitions/User"
      stepId:
        type: string
      parentId:
        type: string
      slug:
        type: string
      full-slug:
        type: string
      posted:
        type: string
        format: date-time
      text:
        type: string